Create README.md
Hello to homeschoolers and those considering that means of educating your children. I am a retired Pediatrician and have been studying the trends in education for the past 50 years. After retirement I began tutoring reading and literacy. I use a very successful program patterned after the Logic of English. If any family is interested , I share my worksheets and especially my approach to to a “crash course” for those children (or adults) who struggle with reading. I do not charge for these services for it is a labor of love and part of my Christian witness